reviver fixing

+++ selection_sort/selection_sort_scripts/selection_sort.h
@@ -3,6 +3,6 @@
 #include <stddef.h>
 #include "sort_work_lib.h"
 
-void selection_sort(void *array, int amount, size_t size, int (*cmp)(const void *a, const void *b));
+void selection_sort(void *array, size_t amount, size_t size, int (*cmp)(const void *a, const void *b));
 #endif
